knowledge-service/requirements.txt
# 知识库服务 Python 依赖包

# FastAPI 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP 请求处理
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1
aiohttp==3.9.1          # 异步HTTP客户端
markitdown==0.0.1a2     # 文档格式化处理

# 数据库相关
sqlalchemy==2.0.23
asyncpg==0.29.0          # PostgreSQL 异步驱动
psycopg2-binary==2.9.9   # PostgreSQL 同步驱动
redis==5.0.1             # Redis 客户端
alembic==1.13.0          # 数据库迁移

# 对象存储
minio==7.2.0             # MinIO 客户端

# 向量数据库（可选，根据配置启用）
# pymilvus==2.3.4         # Milvus 向量数据库
# pgvector==0.2.4         # PostgreSQL 向量扩展
# elasticsearch==8.11.0   # ElasticSearch
# lancedb==0.4.0          # LanceDB

# LlamaIndex 框架
llama-index==0.10.25
llama-index-embeddings-openai==0.1.9
llama-index-embeddings-azure-openai==0.1.9
llama-index-embeddings-huggingface==0.1.9
llama-index-vector-stores-postgres==0.1.9
llama-index-vector-stores-milvus==0.1.9
llama-index-readers-file==0.1.9
llama-index-readers-web==0.1.9

# Agno 框架集成
agno==1.7.0

# 文本处理和嵌入
openai==1.3.7                    # OpenAI API 客户端
tiktoken==0.5.2                  # OpenAI 分词器
sentence-transformers==2.2.2     # HuggingFace 句子嵌入
transformers==4.36.0             # HuggingFace 变换器
torch==2.1.1                     # PyTorch（CPU版本）

# 文档处理
pypdf2==3.0.1           # PDF 处理
python-docx==1.1.0      # DOCX 处理
python-multipart==0.0.6 # 文件上传支持
beautifulsoup4==4.12.2  # HTML 解析
markdown==3.5.1         # Markdown 处理

# 数据处理和工具
pandas==2.1.4           # 数据分析
numpy==1.25.2           # 数值计算
scikit-learn==1.3.2     # 机器学习工具

# 日志和监控
structlog==23.2.0       # 结构化日志
python-json-logger==2.0.7  # JSON 日志格式

# 开发和测试工具
pytest==7.4.3           # 测试框架
pytest-asyncio==0.21.1  # 异步测试支持
black==23.11.0          # 代码格式化
isort==5.12.0           # 导入排序
flake8==6.1.0           # 代码检查

# 异步和并发
anyio==4.1.0            # 异步工具
asyncio-mqtt==0.16.1    # 异步 MQTT
celery==5.3.4           # 分布式任务队列（可选）

# 配置和环境
python-dotenv==1.0.0    # 环境变量加载
dynaconf==3.2.4         # 动态配置管理

# 安全和认证
cryptography==41.0.8    # 加密工具
python-jose[cryptography]==3.3.0  # JWT 处理

# 类型检查
mypy==1.7.1             # 静态类型检查
types-requests==2.31.0.10  # requests 类型注解

# 其他工具
python-dateutil==2.8.2  # 日期处理
pytz==2023.3            # 时区处理
click==8.1.7            # 命令行工具
rich==13.7.0            # 富文本输出 